requirements.txt
Appium-Python-Client
selenium
pytest>=5.0.0
allure-pytest>=2.8.18
